social_media_api/requirements.txt
Django>=5.1.6
djangorestframework>=3.14.0
django-filter>=23.5
mysqlclient>=2.2.3
gunicorn>=21.2.0
whitenoise>=6.6.0
python-dotenv>=1.0.0
dj-database-url>=2.1.0